refactor: make DerEncoder an outer class
