+++ google-http-client/src/main/java/com/google/api/client/json/webtoken/JsonWebSignature.java
@@ -574,35 +574,4 @@ public class JsonWebSignature extends JsonWebToken {
             SecurityUtils.getSha256WithRsaSignatureAlgorithm(), privateKey, contentBytes);
     return content + "." + Base64.encodeBase64URLSafeString(signature);
   }
-
-  static class DerEncoder {
-    private static byte DER_TAG_SIGNATURE_OBJECT = 0x30;
-    private static byte DER_TAG_ASN1_INTEGER = 0x02;
-
-    static byte[] encode(byte[] signature) {
-      // expect the signature to be 64 bytes long
-      Preconditions.checkState(signature.length == 64);
-
-      byte[] int1 = new BigInteger(1, Arrays.copyOfRange(signature, 0, 32)).toByteArray();
-      byte[] int2 = new BigInteger(1, Arrays.copyOfRange(signature, 32, 64)).toByteArray();
-      byte[] der = new byte[6 + int1.length + int2.length];
-
-      // Mark that this is a signature object
-      der[0] = DER_TAG_SIGNATURE_OBJECT;
-      der[1] = (byte) (der.length - 2);
-
-      // Start ASN1 integer and write the first 32 bits
-      der[2] = DER_TAG_ASN1_INTEGER;
-      der[3] = (byte) int1.length;
-      System.arraycopy(int1, 0, der, 4, int1.length);
-
-      // Start ASN1 integer and write the second 32 bits
-      int offset = int1.length + 4;
-      der[offset] = DER_TAG_ASN1_INTEGER;
-      der[offset + 1] = (byte) int2.length;
-      System.arraycopy(int2, 0, der, offset + 2, int2.length);
-
-      return der;
-    }
-  }
 }
